feat(cuda): tsgu::grouped_gemm — segment_mm + gather_mm live
spec/commit.md Phase 3 commit 18 (kernel-commit template T1-T6).

Kernel (cuda/csrc/kernels/grouped_gemm/grouped_gemm.cu): one kernel pair
behind the commit-9 schema. Gather mode (reduce=False) is a 32x32-tiled
GEMM with the gather fused into the prologue via an adaptive per-block
vote — a block whose row-tile shares one group (the segment_mm interior)
stages the b k-tile in shared memory and runs the fully tiled inner
loop; mixed tiles (gather_mm's arbitrary idx, segment boundaries) read b
per-row from global through L2, never materialising a gathered operand
(kernels.md Family 3 "our edge"). Scatter-reduce mode (reduce=True,
gradB) requires idx sorted non-decreasing — each block binary-searches
its group's contiguous row range and accumulates outer products in
ascending order. No atomics anywhere in either mode: deterministic by
construction, no separate path under use_deterministic_algorithms.
The op's Python backward stable-argsorts rows before every reduce=True
call, so arbitrary-idx gather_mm gradients honour the contract;
segment_mm's repeat-interleaved idx is sorted by construction.

Wrappers (T4): segment_mm/gather_mm validate per naming.md's error
template and dispatch to the op; the dgl fast path, the packaging
version gate, and both _legacy_* nested-tensor bodies are deleted
(dependency-light per map.md). Oracle A fix: the frozen legacy
gather_mm fallback allocated its output without dtype (latent f64 bug
shadowed upstream by dgl) — sole documented deviation from f19d7b4.

Gates (T5): opcheck/parity(A+B)/gradcheck/determinism/contract files
added for grouped_gemm; test_indexed_matmul regated to CUDA+backend;
full six-stage gate green; CPU matrix green. NVBench target
bench_grouped_gemm.cu (with an in-binary cuBLAS GemmGroupedBatched
reference row, TF32 off) + op-level benchmarks/bench_grouped_gemm.py.

Benchmark record (RTX A1000, f32, quick subset): vs legacy
nested-tensor path 2.6x (segment fwd) to 5.1x (gather fwd); gather_mm
beats the gather+per-group-cuBLAS loop 2.3x. Correctness bars met.
**cuBLAS parity bar NOT met on the segment shape** (NVBench: 0.39x at
D=64, 0.17x at D=256 vs SgemmGroupedBatched; op-level segment_mm 0.54x
vs a per-segment mm loop) — the 32x32/4-per-thread geometry is
register-tile-bound; perf follow-up commit planned (larger register
tiles + vectorised double-buffered loads), same pattern as the
sddmm/spmm/spsm perf commits. Scaffold rule not used: provenance is
custom for every recorded row.

+++ tests/oracle/indexed_matmul.py
@@ -214,6 +214,12 @@ def gather_mm(a: torch.Tensor, b: torch.Tensor, idx_b: torch.Tensor) -> torch.Te
 
     # Convert back to tensors, again, ideally this would be handled natively with no copy
     ab_segmented = torch.cat(nested_ab.unbind(), dim=0)
-    ab = torch.empty((N, D2), device=torchdevice)
+    # Sole deviation from f19d7b4 (commit 18): the original `torch.empty((N,
+    # D2), device=...)` omitted dtype and so silently allocated float32,
+    # making the index_put below RuntimeError for f64 inputs — a latent bug
+    # never hit upstream (the dgl fast path shadowed this fallback). Fixed
+    # minimally so Oracle A is usable at f64; Oracle B (fp64 dense reference)
+    # independently guards against bugs shared with the oracle.
+    ab = torch.empty((N, D2), device=torchdevice, dtype=a.dtype)
     ab[src_idx_reshuffled] = ab_segmented
     return ab
